✨ Exercise: 10.7 Elevator Events
This exercise seems very weird to me. Not sure if this is really what
they were looking for. My solution is a bit different than theirs - but
without any logic actually happening it is hard to tell what would be
better. I am a bit more explicit with my Floor type with it being a
newtype struct rather than a type alias